+++ crypto/src/lib.rs
@@ -4,7 +4,6 @@ mod cipher_ctx;
 mod bn;
 mod ec;
 
-pub mod aes_gmac_siv;
 pub mod secret;
 pub mod random;
 pub mod hash;
@@ -16,7 +15,27 @@ pub mod salsa;
 pub mod typestate;
 pub mod x25519;
 
-pub mod aes;
+#[cfg(target_os = "macos")]
+pub mod aes_fruity;
+#[cfg(not(target_os = "macos"))]
+pub mod aes_openssl;
+#[cfg(target_os = "macos")]
+pub use aes_fruity as aes;
+#[cfg(not(target_os = "macos"))]
+pub use aes_openssl as aes;
+
+mod aes_tests;
+
+#[cfg(target_os = "macos")]
+pub mod aes_gmac_siv_fruity;
+#[cfg(not(target_os = "macos"))]
+pub mod aes_gmac_siv_openssl;
+#[cfg(target_os = "macos")]
+pub use aes_gmac_siv_fruity as aes_gmac_siv;
+#[cfg(not(target_os = "macos"))]
+pub use aes_gmac_siv_openssl as aes_gmac_siv;
+
+
 
 
 /// This must be called before using any function from this library.
@@ -38,3 +57,4 @@ pub fn secure_eq<A: AsRef<[u8]> + ?Sized, B: AsRef<[u8]> + ?Sized>(a: &A, b: &B)
         false
     }
 }
+pub const ZEROES: [u8; 64] = [0_u8; 64];
